requirements.txt
alembic==1.15.2
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.1.0.post1
Flask==3.0.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
qrcode[pil]==8.1
reportlab==4.3.1
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.0.1
WTForms==3.2.1
gunicorn
requests
Flask-SocketIO==5.3.6
beautifulsoup4==4.12.3
